+++ spec/keka_spec.rb
@@ -213,18 +213,50 @@ RSpec.describe Keka do
     end
   end
 
-  describe '.ok' do
-    it { expect(described_class.ok).to be_ok }
+  describe '.err!' do
+    it 'halts without msg' do
+      expect { described_class.err! }.to raise_error do |error|
+        expect(error.result).not_to be_ok
+        expect(error.result.msg).to be_nil
+      end
+    end
+
+    it 'halts with msg' do
+      expect { described_class.err!('foo') }.to raise_error do |error|
+        expect(error.result).not_to be_ok
+        expect(error.result.msg).to eq('foo')
+      end
+    end
+  end
+
+  describe '.ok!' do
+    it 'halts without msg' do
+      expect { described_class.ok! }.to raise_error do |error|
+        expect(error.result).to be_ok
+        expect(error.result.msg).to be_nil
+      end
+    end
+
+    it 'halts with msg' do
+      expect { described_class.ok!('foo') }.to raise_error do |error|
+        expect(error.result).to be_ok
+        expect(error.result.msg).to eq('foo')
+      end
+    end
+  end
+
+  describe '.ok_result' do
+    it { expect(described_class.ok_result).to be_ok }
     it 'accepts message' do
-      result = described_class.ok('foo')
+      result = described_class.ok_result('foo')
       expect(result.msg).to eq 'foo'
     end
   end
 
-  describe '.err' do
-    it { expect(described_class.err).not_to be_ok }
+  describe '.err_result' do
+    it { expect(described_class.err_result).not_to be_ok }
     it 'accepts message' do
-      result = described_class.err('foo')
+      result = described_class.err_result('foo')
       expect(result.msg).to eq 'foo'
     end
   end